customize headers on all tables
