+++ apps/dashboard/app/javascript/active_jobs.js
@@ -2,7 +2,7 @@
 
 import oboe from 'oboe';
 import { supportPath } from './config.js';
-import { cssBadgeForState, capitalizeFirstLetter } from './utils.js'
+import { cssBadgeForState, capitalizeFirstLetter, customizeTableHeaders } from './utils.js'
 
 window.fetch_table_data = fetch_table_data;
 window.create_datatable = create_datatable;
@@ -165,6 +165,7 @@ function create_datatable(options){
                                     },
         processing: true,           // Add the "processing" while json is being downloaded.
         drawCallback: function(settings){
+          customizeTableHeaders('#job_status_table');
           if(options.drawCallback){
             options.drawCallback(settings);
           }
