+++ apps/dashboard/app/javascript/path_selector/path_selector_data_table.js
@@ -1,5 +1,5 @@
 import { OODAlertError } from '../alert';
-import { ariaNotify, hide, show } from "../utils";
+import { ariaNotify, customizeTableHeaders, hide, show } from "../utils";
 
 export class PathSelectorTable {
   _table = null;
@@ -38,6 +38,7 @@ export class PathSelectorTable {
   }
 
   initDataTable() {
+    const tableId = this.tableId;
     this._table = $(`#${this.tableId}`).DataTable({
       autoWidth: false,
       language: {
@@ -54,6 +55,9 @@ export class PathSelectorTable {
           // don't trigger select checkbox column as select
           // if you need to omit more columns, use a "selectable" class on the columns you want to support selection
           selector: 'td:not(:first-child)'
+      },
+      drawCallback: function() {
+        customizeTableHeaders(`#${tableId}`);
       },
       // https://datatables.net/reference/option/dom
       // dom: '', dataTables_info nowrap
