+++ apps/dashboard/app/javascript/utils.js
@@ -150,6 +150,25 @@ export function pushNotify(message, options = {}) {
   }
 }
 
+// rearrange table header labels so button labels are not part of header
+export function customizeTableHeaders(tableId) {
+  $(`${tableId} th.dt-orderable-asc`).each(function(_index, el) {
+      const sortButton = $(el).find('span.dt-column-order');
+      const ariaLabel = sortButton.attr('aria-label');
+      const labelId = `${el.textContent.replaceAll(' ', '_').toLowerCase()}_label`;
+      var labelSpan = document.getElementById(labelId);
+      if (labelSpan === null) {
+          labelSpan = document.createElement('span');
+          labelSpan.id = labelId;
+          $('#header_labels').append(labelSpan);
+      }
+      labelSpan.textContent = ariaLabel;
+      sortButton.removeAttr('aria-label');
+      sortButton.attr('aria-describedby', labelId);
+      console.log(sortButton[0]);
+  });
+}
+
 // Store a boolean value in localStorage
 export function storeBoolean(key, value) {
   localStorage.setItem(key, value ? 'true' : 'false');
