+++ apps/dashboard/app/views/active_jobs/index.html.erb
@@ -41,6 +41,8 @@
   </div>
   <div class="col-auto col-lg col-md-12 table-responsive">
     <h2>Active Jobs</h2>
+    <div id="header_labels" class="d-none">
+    </div>
     <table id="job_status_table" class="table datatable table-hover" cellspacing="0" width="100%">
       <thead>
       <tr>
