+++ apps/dashboard/app/views/shared/_path_selector_table.html.erb
@@ -59,6 +59,8 @@
                   <span class="sr-only" aria-live="polite">Loading...</span> 
                 </div>
               </div>
+              <div id="header_labels" class="d-none">
+              </div>
               <table class="table table-hover table-condensed d-table w-100" id="<%= table_id %>">
                 <caption>Choose a file or directory to use in the form.</caption>
                 <thead>
